refactor: バルク集計+toTallyをagg/aggregateAll.tsに集約しduckdbBridgeを簡素化
duckdbBridgeからドメインロジック(集計ループ+Tally変換)を分離し、
DB接続提供のみの責務に徹する構造へ変更。
